feat(time_offset): add time offset to make timing within model more flexible and intuitive
When converting model times to a human-interpretable am/pm clock, the fact that the timings within the model used 0 to be equivalent to an unspecified time caused confusion as someone new to the model. Adding a user-configurable offset will make the model more flexible in the case that users from other hospitals wish to use it with their own demand profiles, and also allow for more comprehensive and independent control over the timings that the sdec and ctp run with.